modify precision

+++ server/src/test/java/org/apache/iotdb/db/utils/datastructure/TVListTest.java
@@ -18,9 +18,18 @@
  */
 package org.apache.iotdb.db.utils.datastructure;
 
+import static org.junit.Assert.assertEquals;
+
+import java.io.IOException;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
+import org.apache.iotdb.jdbc.Config;
 import org.apache.iotdb.tsfile.read.TimeValuePair;
 import org.apache.iotdb.tsfile.utils.TsPrimitiveType.TsLong;
 import org.junit.Assert;
@@ -32,11 +41,11 @@ public class TVListTest {
   public void testDoubleTVList1(){
     TVList tvList = new DoubleTVList();
     double j = 1.2345678;
-    for (long i = 0; i < 1000; i++) {
+    for (int i = 0; i < 1000; i++) {
       tvList.putDouble(i, j);
     }
     tvList.sort();
-    for (long i = 0; i < tvList.size; i++) {
+    for (int i = 0; i < tvList.size; i++) {
       Assert.assertEquals(j, tvList.getDouble((int)i));
       Assert.assertEquals(i, tvList.getTime((int)i));
     }
